@@ -1,35 +1,39 @@
 {
-  "_from": "eslint-plugin-react-hooks@^1.6.1",
-  "_id": "eslint-plugin-react-hooks@1.7.0",
+  "_from": "eslint-plugin-react-hooks",
+  "_id": "eslint-plugin-react-hooks@4.0.8",
   "_inBundle": false,
-  "_integrity": "sha512-iXTCFcOmlWvw4+TOE8CLWj6yX1GwzT0Y6cUfHHZqWnSk144VmVIRcVGtUAzrLES7C798lmvnt02C7rxaOX1HNA==",
+  "_integrity": "sha512-6SSb5AiMCPd8FDJrzah+Z4F44P2CdOaK026cXFV+o/xSRzfOiV1FNFeLl2z6xm3yqWOQEZ5OfVgiec90qV2xrQ==",
   "_location": "/eslint-plugin-react-hooks",
   "_phantomChildren": {},
   "_requested": {
-    "type": "range",
+    "type": "tag",
     "registry": true,
-    "raw": "eslint-plugin-react-hooks@^1.6.1",
+    "raw": "eslint-plugin-react-hooks",
     "name": "eslint-plugin-react-hooks",
     "escapedName": "eslint-plugin-react-hooks",
-    "rawSpec": "^1.6.1",
+    "rawSpec": "",
     "saveSpec": null,
-    "fetchSpec": "^1.6.1"
+    "fetchSpec": "latest"
   },
   "_requiredBy": [
-    "/react-scripts"
+    "#DEV:/",
+    "#USER"
   ],
-  "_resolved": "https://registry.npmjs.org/eslint-plugin-react-hooks/-/eslint-plugin-react-hooks-1.7.0.tgz",
-  "_shasum": "6210b6d5a37205f0b92858f895a4e827020a7d04",
-  "_spec": "eslint-plugin-react-hooks@^1.6.1",
-  "_where": "/home/jing/jr-assessment-jing/node_modules/react-scripts",
+  "_resolved": "https://registry.npmjs.org/eslint-plugin-react-hooks/-/eslint-plugin-react-hooks-4.0.8.tgz",
+  "_shasum": "a9b1e3d57475ccd18276882eff3d6cba00da7a56",
+  "_spec": "eslint-plugin-react-hooks",
+  "_where": "/Users/xiongjieyi/repositories/react_hooks",
   "bugs": {
     "url": "https://github.com/facebook/react/issues"
   },
   "bundleDependencies": false,
   "deprecated": false,
   "description": "ESLint rules for React Hooks",
+  "devDependencies": {
+    "@typescript-eslint/parser": "^2.26.0"
+  },
   "engines": {
-    "node": ">=7"
+    "node": ">=10"
   },
   "files": [
     "LICENSE",
@@ -48,12 +52,12 @@
   "license": "MIT",
   "name": "eslint-plugin-react-hooks",
   "peerDependencies": {
-    "eslint": "^3.0.0 || ^4.0.0 || ^5.0.0 || ^6.0.0"
+    "eslint": "^3.0.0 || ^4.0.0 || ^5.0.0 || ^6.0.0 || ^7.0.0"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/facebook/react.git",
     "directory": "packages/eslint-plugin-react-hooks"
   },
-  "version": "1.7.0"
+  "version": "4.0.8"
 }
