@@ -1,45 +1,36 @@
-import React from "react"
+import React, { useState } from "react"
 import AssetsInfo from './components/AssetsInfo'
 import CollectionsInfo from './components/CollectionsInfo'
 import "./styles.css"
 
-export default class App extends React.Component {
-  constructor(props){
-    super(props);
-    this.state={
-      assets: [],
-      masterIdObj:{},  
+
+export default function App(){
+  const [assets, setAssets] = useState([])
+  
+  const [masterIdObj, setMasterIdObj] = useState({})
+
+  const setMasterId = (collectionId, masterId)=> {
+    if (masterIdObj[collectionId]!== masterId){
+      const T={...masterIdObj}
+      T[collectionId]=masterId   
+      setMasterIdObj(T)
     }
-}
-
-  setDataFromCollection = (dataFromCollection)=>{
-    this.setState({
-      assets:dataFromCollection
-    })
   }
+  
+  const curMasterId = assets.length === 0 ? 0 : masterIdObj[assets[0].collectionId] 
 
-  setMasterId = (collectionId, masterId)=> {
-    const T=this.state.masterIdObj
-    T[collectionId]=masterId
-    this.setState({masterIdObj: T})
-  }
+  return(
+    <div className="app">
+      <CollectionsInfo setDataFromCollection={(dataFromCollection)=>{setAssets(dataFromCollection)}}
+        setDefaultMasterId={setMasterId}
+        masterIdObj={masterIdObj}     
+        />
+  
+      <AssetsInfo selectedAssets={assets}
+        setMasterId={setMasterId}
+        masterId={curMasterId}
+        />
+    </div>
+  )
 
-  render(){
-    const curMasterId=this.state.assets.length===0 ? 0 : 
-      this.state.masterIdObj[this.state.assets[0].collectionId]  
-
-    return(
-      <div className="app">
-        <CollectionsInfo setDataFromCollection={this.setDataFromCollection}
-          setDefaultMasterId={this.setMasterId}
-          masterIdObj={this.state.masterIdObj}     
-          />
-   
-        <AssetsInfo selectedAssets={this.state.assets}
-          setMasterId={this.setMasterId}
-          masterId={curMasterId}
-          />
-      </div>
-    )
-  }
 }
